+++ ledger/participant-state/kvutils/src/main/protobuf/com/daml/ledger/participant/state/kvutils/wire/submission.proto
@@ -1,0 +1,34 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+syntax = "proto3";
+package com.daml.ledger.participant.state.kvutils.wire;
+option java_package = "com.daml.ledger.participant.state.kvutils.wire";
+option java_multiple_files = true;
+option csharp_namespace = "Com.Daml.Ledger.Participant.State.KVUtils.Wire";
+
+import "com/daml/ledger/participant/state/kvutils/daml_kvutils.proto";
+
+// A submission to the ledger: a payload and its inputs if any.
+// Produced by [[KeyValueSubmission]].
+// Transformed into `DamlLogEntry` when committed.
+message DamlSubmission {
+  repeated DamlStateKey input_daml_state = 1;
+  oneof payload {
+    DamlTransactionEntry transaction_entry = 2;
+    DamlPackageUploadEntry package_upload_entry = 3;
+    DamlConfigurationSubmission configuration_submission = 4;
+    DamlPartyAllocationEntry party_allocation_entry = 5;
+  }
+  bytes submission_seed = 6;
+}
+
+// A batch of enveloped [[DamlSubmission]]'s.
+message DamlSubmissionBatch {
+  message CorrelatedSubmission {
+    bytes submission = 1;
+    string correlation_id = 2;
+  }
+
+  repeated CorrelatedSubmission submissions = 2;
+}
