+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/Envelope.scala
@@ -18,13 +18,13 @@ object Envelope {
 
   sealed trait Message extends Product with Serializable
 
-  final case class SubmissionMessage(submission: Proto.DamlSubmission) extends Message
+  final case class SubmissionMessage(submission: wire.DamlSubmission) extends Message
 
   final case class LogEntryMessage(logEntry: Proto.DamlLogEntry) extends Message
 
   final case class StateValueMessage(value: Proto.DamlStateValue) extends Message
 
-  final case class SubmissionBatchMessage(value: Proto.DamlSubmissionBatch) extends Message
+  final case class SubmissionBatchMessage(value: wire.DamlSubmissionBatch) extends Message
 
   private val DefaultCompression = true
 
@@ -47,10 +47,10 @@ object Envelope {
         .build
     )
 
-  def enclose(sub: Proto.DamlSubmission): Raw.Envelope =
+  def enclose(sub: wire.DamlSubmission): Raw.Envelope =
     enclose(sub, compression = DefaultCompression)
 
-  def enclose(sub: Proto.DamlSubmission, compression: Boolean): Raw.Envelope =
+  def enclose(sub: wire.DamlSubmission, compression: Boolean): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.SUBMISSION, sub.toByteString, compression)
 
   def enclose(logEntry: Proto.DamlLogEntry): Raw.Envelope =
@@ -65,7 +65,7 @@ object Envelope {
   def enclose(stateValue: Proto.DamlStateValue, compression: Boolean): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.STATE_VALUE, stateValue.toByteString, compression)
 
-  def enclose(batch: Proto.DamlSubmissionBatch): Raw.Envelope =
+  def enclose(batch: wire.DamlSubmissionBatch): Raw.Envelope =
     enclose(Proto.Envelope.MessageKind.SUBMISSION_BATCH, batch.toByteString, compression = false)
 
   def open(envelopeBytes: Raw.Envelope): Either[String, Message] =
@@ -95,13 +95,13 @@ object Envelope {
           parseMessageSafe(() => Proto.DamlLogEntry.parseFrom(uncompressedMessage))
             .map(LogEntryMessage)
         case Proto.Envelope.MessageKind.SUBMISSION =>
-          parseMessageSafe(() => Proto.DamlSubmission.parseFrom(uncompressedMessage))
+          parseMessageSafe(() => wire.DamlSubmission.parseFrom(uncompressedMessage))
             .map(SubmissionMessage)
         case Proto.Envelope.MessageKind.STATE_VALUE =>
           parseMessageSafe(() => Proto.DamlStateValue.parseFrom(uncompressedMessage))
             .map(StateValueMessage)
         case Proto.Envelope.MessageKind.SUBMISSION_BATCH =>
-          parseMessageSafe(() => Proto.DamlSubmissionBatch.parseFrom(uncompressedMessage))
+          parseMessageSafe(() => wire.DamlSubmissionBatch.parseFrom(uncompressedMessage))
             .map(SubmissionBatchMessage)
         case Proto.Envelope.MessageKind.UNRECOGNIZED =>
           Left(s"Unrecognized message kind: ${envelope.getKind}")
@@ -114,13 +114,13 @@ object Envelope {
       case msg => Left(s"Expected log entry, got ${msg.getClass}")
     }
 
-  def openSubmission(envelopeBytes: Raw.Envelope): Either[String, Proto.DamlSubmission] =
+  def openSubmission(envelopeBytes: Raw.Envelope): Either[String, wire.DamlSubmission] =
     open(envelopeBytes).flatMap {
       case SubmissionMessage(entry) => Right(entry)
       case msg => Left(s"Expected submission, got ${msg.getClass}")
     }
 
-  def openSubmission(envelopeBytes: Array[Byte]): Either[String, Proto.DamlSubmission] =
+  def openSubmission(envelopeBytes: Array[Byte]): Either[String, wire.DamlSubmission] =
     open(envelopeBytes).flatMap {
       case SubmissionMessage(entry) => Right(entry)
       case msg => Left(s"Expected submission, got ${msg.getClass}")
