+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/Projections.scala
@@ -5,8 +5,7 @@ package com.daml.ledger.participant.state.kvutils
 
 import com.daml.lf.data.BackStack
 import com.daml.lf.data.Ref.Party
-import com.daml.lf.transaction.BlindingInfo
-import com.daml.lf.transaction.{NodeId, Transaction}
+import com.daml.lf.transaction.{BlindingInfo, NodeId, Transaction}
 
 final case class ProjectionRoots(
     party: Party,
