+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/EnvelopeSpec.scala
@@ -3,6 +3,7 @@
 
 package com.daml.ledger.participant.state.kvutils
 
+import com.daml.ledger.participant.state.kvutils.wire._
 import com.daml.ledger.participant.state.kvutils.{DamlKvutils => Proto}
 import com.google.protobuf.ByteString
 import org.scalatest.matchers.should.Matchers
@@ -12,7 +13,7 @@ class EnvelopeSpec extends AnyWordSpec with Matchers {
   "envelope" should {
 
     "be able to enclose and open" in {
-      val submission = Proto.DamlSubmission.getDefaultInstance
+      val submission = DamlSubmission.getDefaultInstance
 
       Envelope.open(Envelope.enclose(submission)) shouldEqual
         Right(Envelope.SubmissionMessage(submission))
@@ -27,9 +28,9 @@ class EnvelopeSpec extends AnyWordSpec with Matchers {
     }
 
     "be able to enclose and open batch submission batch message" in {
-      val submissionBatch = Proto.DamlSubmissionBatch.newBuilder
+      val submissionBatch = DamlSubmissionBatch.newBuilder
         .addSubmissions(
-          Proto.DamlSubmissionBatch.CorrelatedSubmission.newBuilder
+          DamlSubmissionBatch.CorrelatedSubmission.newBuilder
             .setCorrelationId("anId")
             .setSubmission(ByteString.copyFromUtf8("a submission"))
         )
