+++ ledger/participant-state/kvutils/tools/engine-replay/src/replay/scala/ledger/participant/state/kvutils/tools/engine/replay/Replay.scala
@@ -11,7 +11,8 @@ import com.daml.ledger.participant.state.kvutils.export.{
   ProtobufBasedLedgerDataImporter,
   SubmissionInfo,
 }
-import com.daml.ledger.participant.state.kvutils.{Envelope, Raw, DamlKvutils => Proto}
+import com.daml.ledger.participant.state.kvutils.wire.DamlSubmission
+import com.daml.ledger.participant.state.kvutils.{Envelope, Raw}
 import com.daml.lf.archive.UniversalArchiveDecoder
 import com.daml.lf.crypto
 import com.daml.lf.data._
@@ -122,10 +123,10 @@ private[replay] object Replay {
 
   private[this] def decodeSubmission(
       participantId: Ref.ParticipantId,
-      submission: Proto.DamlSubmission,
+      submission: DamlSubmission,
   ): LazyList[TxEntry] =
     submission.getPayloadCase match {
-      case Proto.DamlSubmission.PayloadCase.TRANSACTION_ENTRY =>
+      case DamlSubmission.PayloadCase.TRANSACTION_ENTRY =>
         val entry = submission.getTransactionEntry
         val tx = TxCoder
           .decodeTransaction(
